requirements.txt
alembic==1.12.0
async-timeout==4.0.3
Babel==2.12.1
blinker==1.6.2
bootstrap4==0.1.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
dominate==2.8.0
elastic-transport==8.4.1
elasticsearch==8.10.1
email-validator==2.0.0.post2
Flask==2.3.3
flask-babel==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
guess-language-spirit==0.5.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
rq==1.15.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==1.26.18
visitor==0.1.3
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
